requirements.txt
fastapi
uvicorn
bs4
anthropic
python-dotenv